修改oss 配置文件

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/vo/tree/DeptTreeReslt.java
@@ -23,7 +23,7 @@ import java.util.List;
 public class DeptTreeReslt extends TreeNode{
 
     @ApiModelProperty(value = "负责人")
-    private Integer principal;
+    private Long principal;
 
     @ApiModelProperty(value = "级别0公司，1部门")
     private Integer level;
